# docs/swagger.yaml
definitions:
  api.ErrorResponse:
    properties:
      msg:
        type: string
    type: object
  blacklist.AddIPToBlacklistRequest:
    properties:
      ip:
        type: string
    type: object
  blacklist.IpResponse:
    properties:
      data:
        items:
          type: string
        type: array
    type: object
  blacklist.Response:
    properties:
      msg:
        type: string
    type: object
  user.ChangePasswordRequest:
    properties:
      oldPassword:
        type: string
      password:
        type: string
      password2:
        type: string
    type: object
  user.CreateUserRequest:
    properties:
      password:
        type: string
      password2:
        type: string
      username:
        type: string
    type: object
  user.CreateUserResponse:
    properties:
      username:
        type: string
    type: object
  user.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  user.LoginResponse:
    properties:
      token:
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
  user.Response:
    properties:
      msg:
        type: string
    type: object
  whitelist.AddIPToWhitelistRequest:
    properties:
      ip:
        type: string
    type: object
  whitelist.IpResponse:
    properties:
      data:
        items:
          type: string
        type: array
    type: object
  whitelist.Response:
    properties:
      msg:
        type: string
    type: object
info:
  contact: {}
  description: Web Application Firewall By Gin
  license: {}
  title: WAF
  version: v1.0
paths:
  /waf/blacklist:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: IP信息
        in: body
        name: AddIPToBlacklistRequest
        required: true
        schema:
          $ref: '#/definitions/blacklist.AddIPToBlacklistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blacklist.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 将IP移除黑名单
      tags:
      - IP黑白名单
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blacklist.IpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 获取黑名单列表
      tags:
      - IP黑白名单
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: IP信息
        in: body
        name: AddIPToBlacklistRequest
        required: true
        schema:
          $ref: '#/definitions/blacklist.AddIPToBlacklistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/blacklist.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 添加IP至黑名单
      tags:
      - IP黑白名单
  /waf/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名、密码
        in: body
        name: LoginRequest
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 用户登录
      tags:
      - Auth
  /waf/user:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: 旧密码、新密码、重复密码
        in: body
        name: ChangePasswordRequest
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 修改密码
      tags:
      - Auth
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户名、密码、重复密码
        in: body
        name: CreateUserRequest
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 创建用户
      tags:
      - Auth
  /waf/whitelist:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: IP信息
        in: body
        name: AddIPToWhitelistRequest
        required: true
        schema:
          $ref: '#/definitions/whitelist.AddIPToWhitelistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/whitelist.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 将IP移除白名单
      tags:
      - IP黑白名单
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/whitelist.IpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 获取白名单列表
      tags:
      - IP黑白名单
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization header
        in: header
        name: Authorization
        required: true
        type: string
      - description: IP信息
        in: body
        name: AddIPToWhitelistRequest
        required: true
        schema:
          $ref: '#/definitions/whitelist.AddIPToWhitelistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/whitelist.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: 添加IP至白名单
      tags:
      - IP黑白名单
swagger: "2.0"